+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-05-30 18:52:56
-// Commit: style(storage): harmonize light and dark theme palette for storage
+// Updated at: 2026-05-31 15:58:53
+// Commit: feat(storage): add bi-directional event bus handler for storage
 
 class storageComponentState {
-  final String commitRef = '1633';
-  final String lastMessage = 'style(storage): harmonize light and dark theme palette for storage';
-  final DateTime timestamp = DateTime.parse('2026-05-30T18:52:56+05:00');
+  final String commitRef = '1653';
+  final String lastMessage = 'feat(storage): add bi-directional event bus handler for storage';
+  final DateTime timestamp = DateTime.parse('2026-05-31T15:58:53+05:00');
 }